add config to limit worlds in which structures will be generated
change generator implementation from ChunkPopulateEvent listener to custom BlockPopulator that is added to worlds on WorldInitEvent
