+++ Plugin/src/main/resources/config.yml
@@ -1,0 +1,3 @@
+enableInWorlds:
+- world
+- world_nether
